requirements.txt
discord.py
pyyaml
matplotlib
tortoise-orm
